Updated Missile/Rammer Follow Behavior
Added a new pair of scripts called HeatSeeker and HeatStopper, which enable and disable following behavior. Requires a rigidbody for the HeatSeeker. HeatStopper goes on a trigger volume attached to the player. Small edit to player, but not important and easily replicatable. 
Updated all enemies to be Trigger colliders, and all projectiles to function OnTriggerEnter instead of OnCollision. OnCollision still disables the projectile.