Added laser beam overload visual
Changed laser beam to use object pooling to save resources on overload. Created beam script to handle displaying beam line renderer. Added overload to PoolUtility to handle objects that only need to be set active/inactive and assigned a parent if necessary. Added score to enemies that did have it.