Bullet speed field added to bandit/minion, spearhead bug fixes.
Added in existing method to set speed of a projectile. So now projectile speed can be set on an individual basis for minions and bandits (used to be universal). Spearhead has also been fixed to correctly damage the player and be destroyed when it impacts. Also added a quick fix to have it go further off screen after reaching the player's last position.