Collider adjusted to fit enemy shape, enemy bullet bug fix,
Collider adjusted for bandit and spearhead, just moved it up a bit on the Y to cover the shape better. Added fix to enemy bullet not spawning in correct rotation and applied changes to layer collision matrix (disabled collisions between default/projectiles and projectiles/projectiles) to fix issue of bullets spinning around for bandit and minion.